fix(schema): .int() для price (+3 regression cases)
Третий integer-validation fix серии (после weekDays/monthDay в 513f6cd).

drizzle column tasks.price = int. Без .int() в Zod schema:
  • price=50.5 (юзер набил «50.5» в input) → MySQL cast в 50 silently
    → расхождение между form-input и БД value
  • price=Infinity (corrupted programmatic state) → balance update'ы
    могли бы стать Infinity, UI показал бы «∞ ₽» или NaN

Number.isInteger automatically блокирует Infinity (isInteger=false).

Regression tests:
  • price=50.5 → ZodError
  • price=Infinity → ZodError
  • price=0/100/999999 → success (positive integers)

(typecheck зелёный, 1258 → 1261 +3)

+++ shared/schema.ts
@@ -175,7 +175,11 @@ export const insertTaskSchema = createInsertSchema(tasks).pick({
   // не будет показываться в свой день. Раньше Zod пропускал.
   monthDay: z.number().int().min(1).max(31).nullable().optional(),
   isRecurring: z.boolean().optional().default(true), // повторяющаяся задача
-  price: z.number().min(0).optional().default(0), // стоимость выполнения в рублях
+  // .int() обязателен: drizzle column int. Без guard'а float (например
+  // юзер ввёл «50.5») попадёт в БД через ZodError-bypass, MySQL cast'нет
+  // в 50 silently — расхождение между ZOD-form-input и БД value. Также
+  // .int() блокирует Infinity (Number.isInteger(Infinity)===false).
+  price: z.number().int().min(0).optional().default(0),
   category: z.string().max(100).nullable().optional(), // категория задачи
   // Описание — DB column TEXT (64KB max). Без zod cap юзер мог
   // отправить 50KB описание; multi-tenant SaaS потенциально
